+++ app/controllers/api/profiles_controller.rb
@@ -230,7 +230,7 @@ class API::ProfilesController < API::ApplicationController
   end
 
   def profile_params
-    params.require(:profile).permit(:username, :bio, :intro, :avatar, :allow_discovery, settings: {})
+    params.require(:profile).permit(:username, :bio, :intro, :avatar, :allow_discovery, :is_caregiver, settings: {})
   end
 
   def profile_public_last_modified(profile)